test(BitcoinExplorers): moved test files
