+++ test/application/explorers/bitcoin/blockexplorer.test.js
@@ -1,15 +1,15 @@
-import * as mockBitpayResponse from '../../mocks/mockBitpayResponse';
-import { generateTransactionDataFromBitpayResponse } from '../../../../../src/explorers/bitcoin/bitpay';
+import * as mockBlockexplorerResponse from '../mocks/mockBlockexplorerResponse';
+import { generateTransactionDataFromBlockexplorerResponse } from '../../../../src/explorers/bitcoin/blockexplorer';
 
-function getMockBitpayResponse () {
-  return JSON.parse(JSON.stringify(mockBitpayResponse));
+function getMockBlockexplorerResponse () {
+  return JSON.parse(JSON.stringify(mockBlockexplorerResponse));
 }
 
-describe('Bitpay Explorer test suite', function () {
+describe('Blockexplorer Explorer test suite', function () {
   let mockResponse;
 
   beforeEach(function () {
-    mockResponse = getMockBitpayResponse();
+    mockResponse = getMockBlockexplorerResponse();
   });
 
   describe('given the transaction has enough confirmations', function () {
@@ -21,7 +21,7 @@ describe('Bitpay Explorer test suite', function () {
     };
 
     it('should return the transaction data', function () {
-      expect(generateTransactionDataFromBitpayResponse(mockResponse)).toEqual(assertionTransactionData);
+      expect(generateTransactionDataFromBlockexplorerResponse(mockResponse)).toEqual(assertionTransactionData);
     });
   });
 
@@ -29,8 +29,8 @@ describe('Bitpay Explorer test suite', function () {
     it('should throw the right error', async function () {
       mockResponse.confirmations = 0;
       expect(() => {
-        generateTransactionDataFromBitpayResponse(mockResponse);
-      }).toThrowError('Number of transaction confirmations were less than the minimum required, according to Bitpay API');
+        generateTransactionDataFromBlockexplorerResponse(mockResponse);
+      }).toThrowError('Number of transaction confirmations were less than the minimum required, according to Blockexplorer API');
     });
   });
 });
